refactor: simplify TasksProvider to pure context shell
Remove TasksAppBoundary render-prop component, inline drag/reorder
logic into App.tsx. TasksProvider receives data via props instead
of owning hooks. Fix setProjects missing from useCallback deps.
