+++ apps/desktop/src/renderer/src/App.tsx
@@ -1,26 +1,34 @@
-import { useState, useCallback, useEffect, useRef } from 'react'
+import { useState, useMemo, useCallback, useEffect, useRef } from 'react'
+import type { DragEndEvent } from '@dnd-kit/core'
+import { arrayMove } from '@dnd-kit/sortable'
 import { useQueryClient } from '@tanstack/react-query'
 import { Loader2 } from '@/lib/icons'
 import { AppSidebar } from '@/components/app-sidebar'
 import { SidebarInset, SidebarProvider } from '@/components/ui/sidebar'
 import { Toaster } from '@/components/ui/sonner'
+import { DragProvider, type DragState } from '@/contexts/drag-context'
+import { DroppedPriorityProvider } from '@/contexts/dropped-priority-context'
 import { AIInlineProvider } from '@/contexts/ai-inline-context'
 import { DayPanelProvider } from '@/contexts/day-panel-context'
 import { SidebarDrillDownProvider } from '@/contexts/sidebar-drill-down'
 import { SelectedFolderProvider } from '@/contexts/selected-folder-context'
 import { GlobalDayPanel } from '@/components/day-panel'
 import { TaskDragOverlay } from '@/components/tasks/drag-drop'
+import { taskViews } from '@/data/tasks-data'
 import { ThemeProvider } from 'next-themes'
 
 // Tab System imports
 import { TabProvider, useTabs } from '@/contexts/tabs'
 import { useTabPersistence, useSessionRestore, STORAGE_KEY } from '@/contexts/tabs/persistence'
+import { TasksProvider } from '@/contexts/tasks'
 import { TabDragProvider, TabErrorBoundary } from '@/components/tabs'
 import { SplitViewContainer } from '@/components/split-view'
 import { ChordIndicator, KeyboardShortcutsDialog } from '@/components/keyboard'
 import {
   useTabKeyboardShortcuts,
   useChordShortcuts,
+  useDragHandlers,
+  useTaskOrder,
   useVault,
   useSettingsShortcut,
   useNewNoteShortcut,
@@ -33,6 +41,7 @@ import { SettingsModalProvider, useSettingsModal } from '@/contexts/settings-mod
 import { SettingsModal } from '@/components/settings-modal'
 import { useFolderViewEvents } from '@/hooks/use-folder-view-events'
 import { useFlushOnQuit } from '@/hooks/use-flush-on-quit'
+import { tasksService } from '@/services/tasks-service'
 import { notesService } from '@/services/notes-service'
 import { VaultOnboarding } from '@/components/vault-onboarding'
 import { FirstRunOnboarding } from '@/components/first-run-onboarding'
@@ -40,7 +49,9 @@ import { useThemeSync } from '@/hooks/use-theme-sync'
 import { useGeneralSettings } from '@/hooks/use-general-settings'
 import { createLogger } from '@/lib/logger'
 import { getStartupTheme, THEME_STORAGE_KEY } from '@/lib/startup-theme'
-import { TasksAppBoundary } from '@/features/tasks/tasks-app-boundary'
+import { useTaskWorkspaceData, useTaskWorkspaceMutations } from '@/features/tasks/use-task-queries'
+import { useTaskUiStore } from '@/features/tasks/use-task-ui-store'
+import { getFilteredTasks } from '@/lib/task-utils'
 
 const log = createLogger('App')
 const startupTheme = getStartupTheme()
@@ -210,36 +221,136 @@ function App(): React.JSX.Element {
   // currentPage is still used for sidebar highlight state
   const [currentPage] = useState<AppPage>('inbox')
 
+  const { tasks, projects } = useTaskWorkspaceData({ enabled: isVaultOpen })
+  const {
+    setProjects,
+    updateTask: handleUpdateTask,
+    deleteTask: handleDeleteTask
+  } = useTaskWorkspaceMutations()
+  const { selectedTaskIds, setSelectedTaskIds: updateSelectedTaskIds } = useTaskUiStore()
+  const selectedTaskIdsRef = useRef(selectedTaskIds)
+
+  useEffect(() => {
+    selectedTaskIdsRef.current = selectedTaskIds
+  }, [selectedTaskIds])
+
+  // Calculate view counts dynamically
+  const viewCounts = useMemo(() => {
+    const counts: Record<string, number> = {}
+    taskViews.forEach((view) => {
+      const filtered = getFilteredTasks(tasks, view.id, 'view', projects)
+      counts[view.id] = filtered.length
+    })
+    return counts
+  }, [tasks, projects])
+
+  // Update project task counts
+  const projectsWithCounts = useMemo(() => {
+    return projects.map((project) => {
+      const projectTasks = tasks.filter((t) => t.projectId === project.id)
+      const incompleteTasks = projectTasks.filter((t) => {
+        const status = project.statuses.find((s) => s.id === t.statusId)
+        return status?.type !== 'done'
+      })
+      return { ...project, taskCount: incompleteTasks.length }
+    })
+  }, [projects, tasks])
+
+  // Task handlers (passed to TasksPage)
+  // Task order persistence hook
+  const taskOrder = useTaskOrder({ persist: true })
+
+  const handleReorder = useCallback(
+    (updates: Record<string, string[] | null>) => {
+      taskOrder.applyOrderUpdates(updates)
+      for (const [, taskIds] of Object.entries(updates)) {
+        if (!taskIds) continue
+        const positions = taskIds.map((_, i) => i)
+        void tasksService.reorder(taskIds, positions)
+      }
+    },
+    [taskOrder]
+  )
+
+  // Use the comprehensive drag handlers hook
+  const { handleDragEnd: taskDragEnd, droppedPriorities } = useDragHandlers({
+    tasks,
+    projects,
+    onUpdateTask: handleUpdateTask,
+    onDeleteTask: handleDeleteTask,
+    onReorder: handleReorder,
+    getOrder: taskOrder.getOrder
+  })
+
+  // Combined drag-drop handler (task operations + project reordering)
+  const handleDragEnd = useCallback(
+    (event: DragEndEvent, dragState: DragState) => {
+      const { active, over } = event
+      if (!over) return
+
+      const activeData = active.data.current
+
+      // Handle project reordering in sidebar (not handled by useDragHandlers)
+      if (activeData?.type === undefined && over.id !== active.id) {
+        const activeIndex = projects.findIndex((p) => p.id === active.id)
+        const overIndex = projects.findIndex((p) => p.id === over.id)
+        if (activeIndex !== -1 && overIndex !== -1) {
+          setProjects((prev) => {
+            const reorderedProjects = arrayMove(prev, activeIndex, overIndex)
+            void tasksService.reorderProjects(
+              reorderedProjects.map((project) => project.id),
+              reorderedProjects.map((_, index) => index)
+            )
+            return reorderedProjects
+          })
+          return
+        }
+      }
+
+      // Delegate all task operations to useDragHandlers
+      taskDragEnd(event, dragState)
+
+      // Clear selection after task drag
+      if (dragState.isDragging) {
+        updateSelectedTaskIds(new Set<string>())
+      }
+    },
+    [projects, setProjects, taskDragEnd, updateSelectedTaskIds]
+  )
+
   // Main content with TabProvider and TasksProvider wrapping everything
   // Wrapped in TabErrorBoundary for graceful error handling
   const mainContent = (
-    <TasksAppBoundary>
-      {({ projects, viewCounts }) => (
-        <TabErrorBoundary
-          onError={(error, errorInfo) => log.error('Critical error:', error, errorInfo)}
-        >
-          <DayPanelProvider>
-            <AIInlineProvider>
-              <TabProvider>
-                <TabPersistenceManager>
-                  <SettingsModalProvider>
-                    <SelectedFolderProvider>
-                      <SidebarDrillDownProvider>
-                        <AppSidebar currentPage={currentPage} viewCounts={viewCounts} />
-                        <SidebarInset className="flex flex-col overflow-hidden">
-                          <AppContent />
-                        </SidebarInset>
-                      </SidebarDrillDownProvider>
-                      <TaskDragOverlay projects={projects} />
-                    </SelectedFolderProvider>
-                  </SettingsModalProvider>
-                </TabPersistenceManager>
-              </TabProvider>
-            </AIInlineProvider>
-          </DayPanelProvider>
-        </TabErrorBoundary>
-      )}
-    </TasksAppBoundary>
+    <TabErrorBoundary
+      onError={(error, errorInfo) => log.error('Critical error:', error, errorInfo)}
+    >
+      <TasksProvider
+        tasks={tasks}
+        projects={projectsWithCounts}
+        getOrderedTasks={taskOrder.getOrderedTasks}
+      >
+        <DayPanelProvider>
+          <AIInlineProvider>
+            <TabProvider>
+              <TabPersistenceManager>
+                <SettingsModalProvider>
+                  <SelectedFolderProvider>
+                    <SidebarDrillDownProvider>
+                      <AppSidebar currentPage={currentPage} viewCounts={viewCounts} />
+                      <SidebarInset className="flex flex-col overflow-hidden">
+                        <AppContent />
+                      </SidebarInset>
+                    </SidebarDrillDownProvider>
+                    {/* Drag Overlay - only for task drag to sidebar */}
+                    <TaskDragOverlay projects={projectsWithCounts} />
+                  </SelectedFolderProvider>
+                </SettingsModalProvider>
+              </TabPersistenceManager>
+            </TabProvider>
+          </AIInlineProvider>
+        </DayPanelProvider>
+      </TasksProvider>
+    </TabErrorBoundary>
   )
 
   if (vaultLoading) {
@@ -275,7 +386,16 @@ function App(): React.JSX.Element {
     >
       <ThemeSyncManager>
         <SidebarProvider key={vaultPath}>
-          {mainContent}
+          <DragProvider
+            tasks={tasks}
+            selectedIds={selectedTaskIds}
+            selectedIdsRef={selectedTaskIdsRef}
+            onDragEnd={(event, state) => void handleDragEnd(event, state)}
+          >
+            <DroppedPriorityProvider value={droppedPriorities}>
+              {mainContent}
+            </DroppedPriorityProvider>
+          </DragProvider>
         </SidebarProvider>
         {/* First-run onboarding overlay — shown until user completes or dismisses */}
         {!generalSettingsLoading && !generalSettings.onboardingCompleted && (
