+++ apps/desktop/src/renderer/src/pages/settings/tasks-section.tsx
@@ -9,7 +9,7 @@ import {
 import { Input } from '@/components/ui/input'
 import { ToggleGroup, ToggleGroupItem } from '@/components/ui/toggle-group'
 import { useTaskPreferences } from '@/hooks/use-task-preferences'
-import { useTaskProjects } from '@/features/tasks/use-task-queries'
+import { useTasksContext } from '@/contexts/tasks'
 import { toast } from 'sonner'
 import {
   SettingsHeader,
@@ -27,7 +27,7 @@ const SORT_OPTIONS = [
 
 export function TasksSettings() {
   const { settings, isLoading, updateSettings } = useTaskPreferences()
-  const { projects } = useTaskProjects()
+  const { projects } = useTasksContext()
 
   const activeProjects = projects.filter((p) => !p.isArchived)
